Address review: memory connector API attach/detach; decouple templates; require MongoService; move memory registry; remove any casts
